Move test code to its own test mod
