+++ src/main/java/net/minecraftforge/common/ForgeMod.java
@@ -19,57 +19,11 @@
 
 package net.minecraftforge.common;
 
-import java.util.List;
-import java.util.Random;
-
-import javax.annotation.Nonnull;
-import javax.annotation.Nullable;
-
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
-import org.apache.logging.log4j.Marker;
-import org.apache.logging.log4j.MarkerManager;
-
-import net.minecraft.block.Block;
-import net.minecraft.block.BlockContainer;
-import net.minecraft.block.material.Material;
-import net.minecraft.block.state.IBlockState;
-import net.minecraft.client.renderer.model.BakedQuad;
-import net.minecraft.client.renderer.model.IBakedModel;
-import net.minecraft.client.renderer.model.ItemOverrideList;
-import net.minecraft.client.renderer.model.ModelResourceLocation;
-import net.minecraft.client.renderer.texture.MissingTextureSprite;
-import net.minecraft.client.renderer.texture.TextureAtlasSprite;
-import net.minecraft.init.Blocks;
-import net.minecraft.nbt.NBTTagCompound;
-import net.minecraft.tileentity.TileEntity;
-import net.minecraft.tileentity.TileEntityType;
-import net.minecraft.util.EnumBlockRenderType;
-import net.minecraft.util.EnumFacing;
-import net.minecraft.util.ITickable;
-import net.minecraft.util.math.BlockPos;
-import net.minecraft.world.IBlockReader;
-import net.minecraft.world.IWorldReader;
-import net.minecraft.world.storage.SaveHandler;
-import net.minecraft.world.storage.WorldInfo;
-import net.minecraftforge.client.event.ModelBakeEvent;
-import net.minecraftforge.client.model.ModelDataManager;
-import net.minecraftforge.client.model.data.IDynamicBakedModel;
-import net.minecraftforge.client.model.data.IModelData;
-import net.minecraftforge.client.model.data.ModelProperty;
-import net.minecraftforge.client.model.data.ModelDataMap;
-import net.minecraftforge.common.model.animation.CapabilityAnimation;
-import net.minecraftforge.energy.CapabilityEnergy;
-import net.minecraftforge.event.RegistryEvent;
 import net.minecraftforge.eventbus.api.IEventBus;
-import net.minecraftforge.fluids.UniversalBucket;
-import net.minecraftforge.fluids.capability.CapabilityFluidHandler;
 import net.minecraftforge.fml.FMLWorldPersistenceHook;
 import net.minecraftforge.fml.ModLoadingContext;
 import net.minecraftforge.fml.VersionChecker;
 import net.minecraftforge.fml.WorldPersistenceHooks;
-import net.minecraftforge.fml.common.Mod;
-import net.minecraftforge.fml.common.gameevent.PlayerEvent;
 import net.minecraftforge.fml.config.ModConfig;
 import net.minecraftforge.fml.event.lifecycle.FMLCommonSetupEvent;
 import net.minecraftforge.fml.event.lifecycle.FMLLoadCompleteEvent;
@@ -78,10 +32,25 @@ import net.minecraftforge.fml.event.lifecycle.InterModProcessEvent;
 import net.minecraftforge.fml.event.server.FMLServerStartingEvent;
 import net.minecraftforge.fml.event.server.FMLServerStoppingEvent;
 import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
-import net.minecraftforge.items.CapabilityItemHandler;
 import net.minecraftforge.server.command.ForgeCommand;
 import net.minecraftforge.versions.forge.ForgeVersion;
 import net.minecraftforge.versions.mcp.MCPVersion;
+
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
+
+import net.minecraft.nbt.NBTTagCompound;
+import net.minecraft.world.storage.SaveHandler;
+import net.minecraft.world.storage.WorldInfo;
+import net.minecraftforge.common.model.animation.CapabilityAnimation;
+import net.minecraftforge.energy.CapabilityEnergy;
+import net.minecraftforge.fluids.capability.CapabilityFluidHandler;
+import net.minecraftforge.items.CapabilityItemHandler;
+import net.minecraftforge.fluids.UniversalBucket;
+import net.minecraftforge.fml.common.Mod;
+import net.minecraftforge.fml.common.gameevent.PlayerEvent;
+import org.apache.logging.log4j.Marker;
+import org.apache.logging.log4j.MarkerManager;
 
 @Mod("forge")
 public class ForgeMod implements WorldPersistenceHooks.WorldPersistenceHook
@@ -118,12 +87,9 @@ public class ForgeMod implements WorldPersistenceHooks.WorldPersistenceHook
         modEventBus.addListener(this::preInit);
         modEventBus.addListener(this::postInit);
         modEventBus.addListener(this::onAvailable);
-        modEventBus.addGenericListener(Block.class, this::registerBlocks);
-        modEventBus.addGenericListener(TileEntityType.class, this::registerTileEntities);
         MinecraftForge.EVENT_BUS.addListener(this::serverStarting);
         MinecraftForge.EVENT_BUS.addListener(this::playerLogin);
         MinecraftForge.EVENT_BUS.addListener(this::serverStopping);
-        MinecraftForge.EVENT_BUS.addListener(this::modelBake);
         ModLoadingContext.get().registerConfig(ModConfig.Type.CLIENT, ForgeConfig.clientSpec);
         ModLoadingContext.get().registerConfig(ModConfig.Type.SERVER, ForgeConfig.serverSpec);
         modEventBus.register(ForgeConfig.class);
@@ -142,116 +108,6 @@ public class ForgeMod implements WorldPersistenceHooks.WorldPersistenceHook
         }
     }
 */
-    
-    private static final ModelProperty<Boolean> MAGIC_PROP = new ModelProperty<Boolean>() {};
-    
-    public void modelBake(ModelBakeEvent event)
-    {
-        final IBakedModel stone = event.getModelRegistry().get(new ModelResourceLocation("minecraft:stone"));
-        final IBakedModel dirt = event.getModelRegistry().get(new ModelResourceLocation("minecraft:dirt"));
-        final IBakedModel old = event.getModelRegistry().get(new ModelResourceLocation("forge:modeltest"));
-        event.getModelRegistry().put(new ModelResourceLocation("forge:modeltest"), new IDynamicBakedModel()
-        {
-            @Override
-            public boolean isGui3d()
-            {
-                return false;
-            }
-            
-            @Override
-            public boolean isBuiltInRenderer()
-            {
-                return false;
-            }
-            
-            @Override
-            public boolean isAmbientOcclusion()
-            {
-                return true;
-            }
-            
-            @Override
-            public List<BakedQuad> getQuads(@Nullable IBlockState state, @Nullable EnumFacing side, Random rand, IModelData modelData)
-            {
-                return modelData.getData(MAGIC_PROP) ? stone.getQuads(state, side, rand, modelData) : dirt.getQuads(state, side, rand, modelData);
-            }
-            
-            @Override
-            public TextureAtlasSprite getParticleTexture()
-            {
-                return MissingTextureSprite.getSprite();
-            }
-            
-            @Override
-            public ItemOverrideList getOverrides()
-            {
-                return null;
-            }
-            
-            @Override
-            @Nonnull
-            public IModelData getModelData(@Nonnull IWorldReader world, @Nonnull BlockPos pos, @Nonnull IBlockState state, @Nonnull IModelData tileData)
-            {
-                if (world.getBlockState(pos.down()).getBlock() == Blocks.AIR)
-                {
-                    tileData.setData(MAGIC_PROP, !tileData.getData(MAGIC_PROP));
-                }
-                return tileData;
-            }
-        });
-    }
-    
-    private static final TileEntityType<Tile> TILE_TYPE = TileEntityType.Builder.create(Tile::new).build(null);
-    private static class Tile extends TileEntity implements ITickable {
-
-        public Tile()
-        {
-            super(TILE_TYPE);
-        }
-        
-        private int counter;
-
-        @Override
-        public void tick()
-        {
-            if (world.isRemote && counter++ == 100)
-            {
-                ModelDataManager.requestModelDataRefresh(this);
-                world.markBlockRangeForRenderUpdate(getPos(), getPos());
-            }
-        }
-        
-        @Override
-        public IModelData getModelData()
-        {
-            return new ModelDataMap.Builder().withInitial(MAGIC_PROP, counter < 100).build();
-        }
-    }
-    
-    public void registerBlocks(RegistryEvent.Register<Block> event)
-    {
-        event.getRegistry().register(new BlockContainer(Block.Properties.create(Material.ROCK))
-        {
-            
-            @Override
-            @Nullable
-            public TileEntity createNewTileEntity(IBlockReader worldIn)
-            {
-                return new Tile();
-            }
-            
-            @Override
-            public EnumBlockRenderType getRenderType(IBlockState state)
-            {
-                return EnumBlockRenderType.MODEL;
-            }
-        }.setRegistryName("forge:modeltest"));
-    }
-    
-    public void registerTileEntities(RegistryEvent.Register<TileEntityType<?>> event)
-    {
-        event.getRegistry().register(TILE_TYPE.setRegistryName("forge:modeltest"));
-    }
 
     public void playerLogin(PlayerEvent.PlayerLoggedInEvent event)
     {
